EMP balance (#162)
* initial stuff

* both cases of ear emps should be rebalanced now

* significant changes

removed baseline 25% chance to emp organs
arm guns no longer explode but simply set you on fire
revivers no longer cause heart attacks
brain implants cause knockdown rather than stun
breathing tube implants cause slightly more breath loss on a small probability
eyes take less damage and are less likely to be effected by EMPs
EMPs no longer cause heart attacks on cybernetic hearts

* telegraphing sounds

+++ code/modules/surgery/organs/liver.dm
@@ -265,8 +265,10 @@
 	. = ..()
 	if(. & EMP_PROTECT_SELF)
 		return
-	if(!COOLDOWN_FINISHED(src, severe_cooldown)) //So we cant just spam emp to kill people.
-		owner.adjustToxLoss(10)
-		COOLDOWN_START(src, severe_cooldown, 10 SECONDS)
-	if(prob(emp_vulnerability/severity)) //Chance of permanent effects
-		organ_flags |= ORGAN_SYNTHETIC_EMP //Starts organ faliure - gonna need replacing soon.
+	if(prob(25 / severity))
+		if(!COOLDOWN_FINISHED(src, severe_cooldown)) //So we cant just spam emp to kill people.
+			playsound(owner, 'sound/weapons/ionrifle.ogg', 50, TRUE, -1)
+			owner.adjustToxLoss(10)
+			COOLDOWN_START(src, severe_cooldown, 10 SECONDS)
+		if(prob(emp_vulnerability/severity)) //Chance of permanent effects
+			organ_flags |= ORGAN_SYNTHETIC_EMP //Starts organ faliure - gonna need replacing soon.
